Extract model structs to separate file
Move all model structs (Names, Enterprise, City, Country, etc.) and
their HasData methods from reader.go to a new models.go file. This
improves code organization by separating data models from the Reader
implementation.
